Reorganization of Method headers
Each core now has its own method directory, this makes it a lot easier and cleaner to both maintain and add new methods

When adding new methods, they should be defined in both the LuaFunctions.cpp and the respective Method.h file. Please also add the registry entry to the other cores' method registry, so we know what's implemented or not.

When a method is not implemented, it should point to a nullptr in the registry.